requirements.txt
Flask
Flask_SQLAlchemy
requests
pg8000
